ml-service/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# ML/DL frameworks
tensorflow==2.15.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.0

# Technical indicators
# Note: TA-Lib requires manual C library installation
# For Docker, we install via apk in Dockerfile
# ta-lib==0.4.28

# NLP and transformers
transformers==4.36.2
torch==2.1.2
tokenizers==0.15.1
sentencepiece==0.1.99

# Text processing
nltk==3.8.1

# Utilities
redis==5.0.1
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1

# Logging
loguru==0.7.2

# Data visualization (for notebooks)
matplotlib==3.8.2
seaborn==0.13.1

# Job scheduling
apscheduler==3.10.4

# Testing
pytest==8.0.0
pytest-asyncio==0.23.2
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
